% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coretta2018itapol.R
\docType{package}
\name{coretta2018itapol}
\alias{coretta2018itapol}
\alias{coretta2018itapol-package}
\title{coretta2018itapol: Data package of an acoustic and articulatory study of the
voicing effect in Italian and Polish}
\description{
This package provides datasets containing variuos acoustic and articulatory
measurements of Italian and Polish. The data are part of an exploratory study
on the effect of consonant voicing on preceding vowel duration (the voicing
effect).
}
