% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlc_voff}
\alias{dlc_voff}
\title{DLC tongue knots data of Italian and Polish at vowel offset}
\format{
A tibble with DLC tongue knots data.:
\describe{
 \item{\code{participant}}{Participant's ID.}
 \item{\code{date}}{Date of recording.}
 \item{\code{time_rec}}{Time within recording.}
 \item{\code{time_annot}}{Time within annotation interval.}
 \item{\code{prompt}}{Prompt sentence.}
 \item{\code{annotation}}{Label of annotation interval.}
 \item{\code{knot}}{Knot number.}
 \item{\code{dlc}}{DLC spline.}
 \item{\code{X}}{X coordinate.}
 \item{\code{Y}}{Y coordinate.}
 \item{\code{knot_label}}{Knot label.}
 \item{\code{word}}{Target word.}
 \item{\code{item}}{word ID number}
 \item{\code{ipa}}{IPA transcription of the word}
 \item{\code{c1}}{first consonant (C1)}
 \item{\code{c1_phonation}}{voicing of C1 (\code{voiceless}, \code{voiced})}
 \item{\code{vowel}}{V1 and V2 (\code{a}, \code{o}, \code{u})}
 \item{\code{anteropost}}{backness of the vowel (\code{back}, \code{central})}
 \item{\code{height}}{height of the vowel (\code{high}, \code{mid}, \code{low})}
 \item{\code{c2}}{second consonant (C2)}
 \item{\code{c2_phonation}}{voicing of C2 (\code{voiceless} or \code{voiced})}
 \item{\code{c2_place}}{place of C2 (\code{coronal}, \code{velar})}
}
}
\usage{
dlc_voff
}
\description{
It contains the DLC tongue knots coordinates at the offset of the first vowel
in CVCV target words. The coordinates have been rotated based on a fiducial
line corresponding to the lower surface of a bite plate.
}
\keyword{datasets}
