% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tongue_contours}
\alias{tongue_contours}
\title{Tongue contours in Italian and Polish}
\format{A tibble with 319116 rows and 30 variables:
\describe{
  \item{\code{speaker}}{speaker's ID}
  \item{\code{seconds}}{(s) time within audio file}
  \item{\code{rec_date}}{date and time of recording}
  \item{\code{prompt}}{sentence stimulus}
  \item{\code{label}}{\code{AAA} annotation label}
  \item{\code{TT_displacement_sm}}{smoothed Tongue Tip displacement}
  \item{\code{TT_velocity}}{Tongue Tip velocity}
  \item{\code{TT_velocity_abs}}{absolute Tongue Tip velocity}
  \item{\code{TD_displacement_sm}}{smoothed Tongue Dorsum displacement}
  \item{\code{TD_velocity}}{Tongue Dorsum velocity}
  \item{\code{TD_velocity_abs}}{absolute Tongue Dorsum velocity}
  \item{\code{TR_displacement_sm}}{smoothed Tongue Root displacement}
  \item{\code{TR_velocity}}{Tongue Root velocity}
  \item{\code{TR_velocity_abs}}{Absolute Tongue Root velocity}
  \item{\code{fan_line}}{number of fan line}
  \item{\code{X}}{x coordinate}
  \item{\code{Y}}{y coordinate}
  \item{\code{word}}{word stimulus}
  \item{\code{language}}{speaker's native language (\code{Italian}, \code{Polish})}
  \item{\code{sex}}{speaker's sex (\code{f}, \code{m})}
  \item{\code{item}}{word ID number}
  \item{\code{ipa}}{IPA transcription of the word}
  \item{\code{c1}}{first consonant (C1)}
  \item{\code{c1_phonation}}{voicing of C1 (\code{voiceless}, \code{voiced})}
  \item{\code{vowel}}{V1 and V2 (\code{a}, \code{o}, \code{u})}
  \item{\code{anteropost}}{backness of the vowel (\code{back}, \code{central})}
  \item{\code{height}}{height of the vowel (\code{high}, \code{mid}, \code{low})}
  \item{\code{c2}}{second consonant (C2)}
  \item{\code{c2_phonation}}{voicing of C2 (\code{voiceless} or \code{voiced})}
  \item{\code{c2_place}}{place of C2 (\code{coronal}, \code{velar})}
}}
\usage{
tongue_contours
}
\description{
Coordinates of tongue contour splines of Italian and Polish at defined time points.
}
\keyword{datasets}
